Initial commit of templates and basic Django folder layout
Not really happy with any of this code at this point but in order to prevent more requests to push things, here it is. Remember to get Compass installed and working if you want to edit (or even view!) the css code (see the readme.md file in the css/ folder for an explanation as to why I've not made it easy). I've set the bar low, let's go. Relates to issue #6, FWIW.